Added the feature to refresh token.
Created a rest endpoint /api/security/refreshToken to refresh if the life of the token is about to expire in 10 (default) minutes. This can be configured with the property system wide "jwt.token.default.refresh". The response of the refresh token is same as that of verifyToken and authenticate end points.