File size issue is fixed
Fixed the buffer size issue. Now can 50 mb file can also be downloaded. Fixed issue with the file path to store the file in a location. Also if no filename is provided, name is taken from the url by stripping everything before the last slash and everything from '?' if there are any parameters.