updated AbstractFilesResourceService
While importing the zip file when the file size is huge it is taking a lot of time for importing because for every small file we are calling file access service method multiple times in turn it will query database each time. So instead of calling each file instead checking for the folder which will reduce the queries called.