Patch field changes
When any model related to security server was invoked with patch request to update any field which belongs to that model. This changes were handled by using reflection concept to update eligible fields by fetching the method related to provided field and iterating in all declared methods of that object. Once the object was fetched invoking the method with obtained object from read by id and updating that field. At last sending the updated object to put method.