Fix bug in package and role creation
The bug caused not able to add same name packages/roles even when apps are different. In this fix we removed unique key constraint from name and added unique key constraint for combination of name and app_id. Also added mandatory check for existence of app_id if client_type is not system.