Finished witht he services changes and dao changes.
Next move is to add reg objects to the newly created user and client.

File access and app access are already added to the client. So did the user role to the user.

Remaining are departments, designations, profile restrictions at client level and user designation, user profile at user level.